backend/requirements.txt
# DeepAgents Framework (latest available version)
deepagents>=0.0.11rc1

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# LangChain & LLMs
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0

# Database (optional for sessions)
aiosqlite>=0.19.0
